% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bark.R
\name{get_bark_key}
\alias{get_bark_key}
\title{Get Current Bark Device Key}
\usage{
get_bark_key()
}
\value{
The currently set Bark device key
}
\description{
Returns the currently set default Bark device key
}
\examples{
\dontrun{
get_bark_key()
}
}
