% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bark.R
\name{set_bark_key}
\alias{set_bark_key}
\title{Set Default Bark Device Key}
\usage{
set_bark_key(device_key)
}
\arguments{
\item{device_key}{Your Bark device key (full URL or key part)}
}
\description{
Sets the default Bark device key to be used by \code{bark_notify()}
}
\examples{
\dontrun{
set_bark_key("your_device_key")
}
}
